add method to generate ternary polynomial
the method for generating random polynomials is important. in particular, if the coefficients are not sufficiently small, the hom_prod_test fails. in accordance with Katherine Stange's notes, it makes more sense to just use ternary polynomials for each "small" polynomial. using a normal distribution works for some, but not all choices of parameters.